Move ExtractIntentAsync from IntentExtractor to DialogRouter
This is a breaking change.

Create new class that inherits `DialogRouterBase` and move your existing `ExtractIntentAsync` in `IntentExtractor` to the new class. Then remove IntentExtractor from the 3D model and attach DialogRouter instead.
