+++ ChatdollKit/Scripts/Dialog/IDialogRouter.cs
@@ -1,13 +1,14 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 
-
 namespace ChatdollKit.Dialog
 {
-    public interface IIntentExtractor
+    public interface IDialogRouter
     {
         void Configure();
+        void RegisterIntent(string intentName, IDialogProcessor dialogProcessor);
         Task<Response> ExtractIntentAsync(Request request, Context context, CancellationToken token);
         Task ShowResponseAsync(Response response, Request request, Context context, CancellationToken token);
+        IDialogProcessor Route(Request request, Context context, CancellationToken token);
     }
 }
